+++ app/models/order.rb
@@ -2,6 +2,8 @@ class Order < ApplicationRecord
   belongs_to :user
   belongs_to :item
   has_one :residence
+  attr_accessor :token
+  
   
 end
